initial commit for chess project

+++ src/main/java/com/tiy/ssa/weekthreeShape/Shape.java
@@ -4,7 +4,7 @@ public interface Shape {
 
     public String word = "word";
     
-    public int getArea();
+    public float getArea();
     
     public void draw();
 }
